# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  catalog.UpsertMedicineInput:
    properties:
      atc_code:
        type: string
      brand_name:
        type: string
      form:
        type: string
      generic_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      manufacturer_id:
        type: string
      pack_size:
        type: string
      strength:
        type: string
    type: object
...
  response.ErrorEnvelope:
    properties:
      error:
        properties:
          code:
            type: string
          details: {}
          message:
            type: string
        type: object
    type: object
...
info:
  contact: {}
  description: Backend API for PharmaLink B2B marketplace
  title: PharmaLink API
  version: "1.0"
paths:
  /admin/medicines/:
    post:
      consumes:
      - application/json
      parameters:
      - description: medicine payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/catalog.UpsertMedicineInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create medicine (Admin)
      tags:
      - catalog
  /admin/medicines/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: medicine id
        in: path
        name: id
        required: true
        type: string
      - description: medicine payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/catalog.UpsertMedicineInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update medicine (Admin)
      tags:
      - catalog
...
  /discount-campaigns/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List discount campaigns (Wholesaler)
      tags:
      - discounts
    post:
      consumes:
      - application/json
      parameters:
      - description: campaign payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create discount campaign (Wholesaler)
      tags:
      - discounts
  /discount-campaigns/{id}/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: campaign id
        in: path
        name: id
        required: true
        type: string
      - description: discount item payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add item to discount campaign (Wholesaler)
      tags:
      - discounts
  /inventory/movements:
    post:
      consumes:
      - application/json
      parameters:
      - description: movement payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add inventory movement (Wholesaler)
      tags:
      - inventory
  /inventory/stock:
    get:
      parameters:
      - description: medicine id
        in: query
        name: medicine_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get current stock by medicine (Wholesaler)
      tags:
      - inventory
...
  /medicines:
    get:
      parameters:
      - description: search query
        in: query
        name: query
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: List/search medicines
      tags:
      - catalog
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"